Initial functionality complete
Created a meteor application with the initial functionality for the timer application. The application currently can play/ pause, stop and record laps. Next things to come styling the application, caching the state so that the application can continue to run on refresh and adding a modal to print out the stats from the timer being run.